chore: Rename mappers to be like checkbox options on export form.
