+++ feet/src/mappers/fire-loop-utils.ts
