+++ feet/src/mappers/fire-panel-utils.ts
