+++ feet/src/pages/feetpage/ExportForm.tsx
@@ -10,13 +10,13 @@ import { useToast } from '../../utils/useToast';
 import { useDataContext } from '../../utils/DataProvider';
 import InfoBox from '../../components/InfoBox';
 import GenericButton from '../../components/GenericButton';
-import { mapPanelToExcel } from '../../mappers/panel-utils';
+import { mapPanelToExcel } from '../../mappers/fire-panel-utils';
 import { mapPanelsWithZones } from '../../mappers/zone-utils';
-import { mapLoopToExcel } from '../../mappers/loop-utils';
-import { mapBoardToExcel } from '../../mappers/board-utils';
-import { mapLoopAddressToExcel } from '../../mappers/address-utils';
+import { mapLoopToExcel } from '../../mappers/fire-loop-utils';
+import { mapBoardToExcel } from '../../mappers/io-board-utils';
+import { mapLoopAddressToExcel } from '../../mappers/address-report-utils';
 import { mapToIOReportToExcel } from '../../mappers/io-report-utils';
-import { mapControlGroupsToExcel } from '../../mappers/controlgroup-utils';
+import { mapControlGroupsToExcel } from '../../mappers/control-group-report-utils';
 import { addSheetToWorkbook, feetLanguages } from '../../mappers/utils';
 import styles from './ExportForm.module.less';
